+++ packages/apps/explorer/src/components/block-table/block-row/block-row.css.ts
@@ -1,0 +1,29 @@
+import { atoms } from '@kadena/react-ui/styles';
+import { style } from '@vanilla-extract/css';
+
+export const textStyle = style([atoms({ alignItems: 'center' })]);
+
+export const rowLinkElementStyle = style([
+  atoms({
+    justifyContent: 'center',
+    padding: 'sm',
+    width: '100%',
+  }),
+]);
+
+export const rowTextElementStyle = style([
+  atoms({
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 'sm',
+    width: '100%',
+  }),
+]);
+
+export const rowChartElementStyle = style([
+  atoms({
+    justifyContent: 'center',
+    alignItems: 'flex-start',
+    paddingBlockStart: 'sm',
+  }),
+]);