+++ packages/apps/explorer/src/services/block.ts
@@ -1,0 +1,62 @@
+import type {
+  BlocksFromHeightsQueryResult,
+  NewBlocksSubscriptionResult,
+} from '@/__generated__/sdk';
+
+interface IBlockData {
+  hash: string;
+  height: number;
+  chainId: number;
+  difficulty: string;
+  txCount: number;
+}
+
+export interface IHeightBlock {
+  [height: number]: IBlockData;
+}
+
+export interface IChainBlock {
+  [chainId: number]: IHeightBlock;
+}
+
+export function addBlockData(
+  existingData: IChainBlock,
+  newData:
+    | NewBlocksSubscriptionResult['data']
+    | BlocksFromHeightsQueryResult['data'],
+): IChainBlock {
+  const updatedData = { ...existingData };
+
+  if (!newData) return updatedData;
+
+  const addBlock = (block: any) => {
+    if (!updatedData[block.chainId]) {
+      updatedData[block.chainId] = {};
+    }
+
+    updatedData[block.chainId][block.height] = {
+      ...block,
+      txCount: block.transactions.totalCount,
+    };
+  };
+
+  if ('newBlocks' in newData) {
+    if (!newData.newBlocks) return updatedData;
+
+    for (const block of newData.newBlocks) {
+      addBlock(block);
+    }
+  } else if ('blocksFromHeight' in newData) {
+    if (!newData.blocksFromHeight) return updatedData;
+
+    for (const block of newData.blocksFromHeight.edges) {
+      if (!block?.node) {
+        continue;
+      }
+
+      addBlock(block.node);
+    }
+  }
+
+  return updatedData;
+}