+++ packages/libs/toolbox/tests/test.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import {} from '../src/config';
-
-describe('test', () => {
-  it.todo('pass', () => {
-    expect(true).toBe(true);
-  });
-});